third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/shadow/elements-panel-shadow-selection-on-refresh-3-expected.txt

Tests that elements panel preserves selected shadow DOM node on page refresh.


Running: setup

Running: testClosedShadowRootChild
Page reloaded.
Selected node: 'button'

Running: testUserAgentShadowRootChild
Page reloaded.
Selected node: 'div'
